requirements.txt
streamlit
pandas
base58
numpy
pillow
plotly
scikit-learn
lazypredict
seaborn
matplotlib
xgboost
lightgbm
pytest
tqdm
ydata-profiling
streamlit-pandas-profiling
Jinja2
